+++ djangocms_frontend/contrib/contact_form/cms_plugins.py
@@ -1,5 +1,0 @@
-from djangocms_frontend.contrib.forms import register_with_frontend
-
-from .forms import ContactForm
-
-register_with_frontend(ContactForm)