+++ djangocms_frontend/contrib/frontend_forms/urls.py
@@ -1,4 +1,4 @@
-from django.urls import include, path
+from django.urls import path
 
 from . import views
 
@@ -17,6 +17,3 @@ urlpatterns = [
         name="ajaxview",
     ),
 ]
-
-
-urls = include("djangocms_frontend.contrib.forms.urls", namespace="dcf_forms")