+++ djangocms_frontend/contrib/link/urls.py
@@ -1,4 +1,4 @@
-from django.urls import include, path
+from django.urls import path
 
 from . import views
 
@@ -7,5 +7,3 @@ app_name = "djangocms_frontend"
 urlpatterns = [
     path("link:autocomplete", views.AutocompleteJsonView.as_view(), name="ac_view"),
 ]
-
-urls = include("djangocms_frontend.contrib.link.urls", namespace="dcf_autocomplete")